% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_lv.R
\name{print.lv}
\alias{print.lv}
\title{Display Letter Values.}
\usage{
\method{print}{lv}(x, ...)
}
\arguments{
\item{x}{an object \code{lv}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
No return value, called to improve visualization of letter values as proposed at the seminal book Understanding Robust and Exploratory Data Analysis by Hoaglin, Mosteller and Tukey published of 1983.
}
\description{
Display and return letter values.
}
\details{
The diagram is, by default, the 5-number summary, where the sample size, the median (location measure) and the F-spread (distance between lower fourth and upper fourth). Others diagrams are avalaible increasing the argument \code{level}.

This function computes the letter values as presented at Understanding Robust and Exploratory Data Analysis by Hoaglin, Mosteller and Tukey published in 1983.

This is a generic print method for the class "lv".
}
\examples{
lv_obj <- letter_value(rivers)
print.lv(lv_obj)

}
