% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lv.R
\name{summary.lv}
\alias{summary.lv}
\title{Summary Using Letter Value}
\usage{
\method{summary}{lv}(object, ...)
}
\arguments{
\item{object}{an object \code{lv}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Compute the resume measures (location and scale) using letter values.
}
\details{
In this summary, we present the trimean, median, F-spread, F-pseudo sigma, F-pseudo variance e outliers values.

This function returns the measures of location and scale as presented at Understanding Robust and Exploratory Data Analysis by Hoaglin, Mosteller and Tukey published in 1983.

This is a generic method for the class "lv".
}
\examples{
lv_obj <- letter_value(rivers)
summary.lv(lv_obj)

}
